Organize headers My first attempt on ed25519 was an unholy mess. It was very disorganized and hard to follow.
