+++ donna_64.cpp
@@ -18,11 +18,7 @@ extern const char DONNA64_FNAME[] = __FILE__;
 
 #if defined(CRYPTOPP_CURVE25519_64BIT)
 
-#if defined(_MSC_VER)
-# include <intrin.h>
-# pragma intrinsic(_umul128)
-# pragma intrinsic(__shiftright128)
-#endif
+#include "donna_64.h"
 
 ANONYMOUS_NAMESPACE_BEGIN
 
@@ -36,51 +32,8 @@ using CryptoPP::sword64;
 using CryptoPP::GetBlock;
 using CryptoPP::LittleEndian;
 
-typedef word64 bignum25519[5];
-
-#if defined(CRYPTOPP_WORD128_AVAILABLE)
-using CryptoPP::word128;
-# define lo128(a) ((word64)a)
-# define hi128(a) ((word64)(a >> 64))
-# define add128(a,b) a += b;
-# define add128_64(a,b) a += (word64)b;
-# define mul64x64_128(out,a,b) out = (word128)a * b;
-# define shr128(out,in,shift) out = (word64)(in >> (shift));
-// # define shl128(out,in,shift) out = (word64)((in << shift) >> 64);
-
-#elif defined(_MSC_VER)
-struct word128 { word64 lo, hi; };
-# define mul64x64_128(out,a,b) out.lo = _umul128(a,b,&out.hi);
-# define shr128_pair(out,hi,lo,shift) out = __shiftright128(lo, hi, shift);
-// # define shl128_pair(out,hi,lo,shift) out = __shiftleft128(lo, hi, shift);
-# define shr128(out,in,shift) shr128_pair(out, in.hi, in.lo, shift)
-// # define shl128(out,in,shift) shl128_pair(out, in.hi, in.lo, shift)
-# define add128(a,b) { word64 p = a.lo; a.lo += b.lo; a.hi += b.hi + (a.lo < p); }
-# define add128_64(a,b) { word64 p = a.lo; a.lo += b; a.hi += (a.lo < p); }
-# define lo128(a) (a.lo)
-# define hi128(a) (a.hi)
-
-#elif defined(__GNUC__)
-struct word128 { word64 lo, hi; };
-# define mul64x64_128(out,a,b) __asm__ ("mulq %3" : "=a" (out.lo), "=d" (out.hi) : "a" (a), "rm" (b));
-# define shr128_pair(out,hi,lo,shift) __asm__ ("shrdq %2,%1,%0" : "+r" (lo) : "r" (hi), "J" (shift)); out = lo;
-// # define shl128_pair(out,hi,lo,shift) __asm__ ("shldq %2,%1,%0" : "+r" (hi) : "r" (lo), "J" (shift)); out = hi;
-# define shr128(out,in,shift) shr128_pair(out,in.hi, in.lo, shift)
-// # define shl128(out,in,shift) shl128_pair(out,in.hi, in.lo, shift)
-# define add128(a,b) __asm__ ("addq %4,%2; adcq %5,%3" : "=r" (a.hi), "=r" (a.lo) : "1" (a.lo), "0" (a.hi), "rm" (b.lo), "rm" (b.hi) : "cc");
-# define add128_64(a,b) __asm__ ("addq %4,%2; adcq $0,%3" : "=r" (a.hi), "=r" (a.lo) : "1" (a.lo), "0" (a.hi), "rm" (b) : "cc");
-# define lo128(a) (a.lo)
-# define hi128(a) (a.hi)
-#else
-// https://groups.google.com/forum/#!forum/cryptopp-users
-# error "Unsupported platform"
-#endif
-
-#define ALIGN(n) CRYPTOPP_ALIGN_DATA(n)
-
-const byte basePoint[32] = {9};
-const word64 reduce_mask_51 = ((word64)1 << 51) - 1;
-// const word64 reduce_mask_52 = ((word64)1 << 52) - 1;
+// Bring in all the symbols from the 64-bit header
+using namespace CryptoPP::Donna::Donna64;
 
 /* out = in */
 inline void
